Add two implementatons for parsing floating-pointer literals
1. "impl1": toy-graded, very imprecise; able to handle
   scientific notation, resort to strtod() for "nasty" situations.
   enabled by adding -DFP_RELAX=2 to CFLAGS.

2. "impl2", almost strict, is able to parse floating point
  in this format quickly: "n...n <decimal-point> m...m" where
  the interger part contains no more than 20 digit, and fraction
  part contains no more than 16 digits. But the parsing result
  is not 100% precise, it would incur up to 1/(2**53) relative
  error, which is nothing for most situations.

  It is enabled by adding -DFP_RELAX=1 to FLAGS

The original implemention can precisely parse floating point
literal (so long as strtod() is precise), and can be turned
on by -DFP_RELEAX=0, which is default for now.

 To reason to keep "impl1" is to set a bar (in terms of parsing
speed) for future work. "Impl2" is useful so long as the
applications are not super picky about floating point precision.

Comparing the 3 implemtatons with citylots.json, we have
 - original (-DFP_RELAX=0), 2.49s
 - impl1", (-DFP_RELAX=2), 1.7s
 - impl2", (-DFP_RELAX=2), 1.6s
 - cjson (not lua-resty-libcjson), 2.35s (The way cjson parse
   floating pointer does not seems to be correct either)

This change is not fully tested yet, and my holy grail for parsing
floating point literl is:
http://www.exploringbinary.com/correct-decimal-to-floating-point-using-big-integers/