Fix a build problem (OSX).
On linux, "uint" seems to be defined in stdlib.h, that is why we don't have build problem on Linux.

 The relavant code is excerpted as following:
 46     const char* np = "-0123456789";
 ..
 48     for (idx = 0, idx_e = strlen(np); idx < idx_e; idx++) {
 49         token_predict[(uint)np[idx]] = TT_FP;
 50     }

The purpose of casting expression "np[idx]" to unsigned integer is to make sure the expression
hold a possitive value of [0..255] (just in case in the future we add a char to "np" with
value >= 128).

Casting the expression to uint (actually uint32_t) is *WRONG*! We should cast the expression
to uint8_t.